+++ src/extensions/ControllerExtension.php
@@ -8,7 +8,8 @@ use SilverStripe\Core\Extension;
  * Extension for Controller that provide methods such as member management
  * interface to templates
  *
- * @package users
+ * @package Users
+ * @author  i-lateral <info@ilateral.co.uk>
  */
 class ControllerExtension extends Extension
 {